Initial commit of Bittrex notifications

+++ TeleCoinigy/Helpers/Helpers.cs
@@ -6,25 +6,25 @@ namespace TeleCoinigy.Helpers
 {
     public static class Helpers
     {
-        public static double BalanceForAuthId(JObject jObject)
+        public static decimal BalanceForAuthId(JObject jObject)
         {
             var data = jObject["data"];
-            return data.Select(token => token["btc_balance"].ToString()).Aggregate<string, double>(0, (current, btcBalance) => current + double.Parse(btcBalance));
+            return data.Select(token => token["btc_balance"].ToString()).Aggregate<string, decimal>(0, (current, btcBalance) => current + decimal.Parse(btcBalance));
         }
 
-        public static double GetLastBid(JObject jObject)
+        public static decimal GetLastBid(JObject jObject)
         {
             var data = jObject["data"];
             var stringBid = data[0]["bid"].ToString();
-            return double.Parse(stringBid);
+            return decimal.Parse(stringBid);
         }
 
-        public static double TotalBtcBalance(JObject jObject)
+        public static decimal TotalBtcBalance(JObject jObject)
         {
-            double btcBalance = 0;
+            decimal btcBalance = 0;
             btcBalance = jObject["data"]
                 .Select(token => token["btc_balance"].ToString())
-                .Select(double.Parse)
+                .Select(decimal.Parse)
                 .Aggregate(btcBalance, (current, d) => current + d);
             return btcBalance;
         }
