+++ TeleCoinigy/Models/BalanceHistory.cs
@@ -4,9 +4,9 @@ namespace TeleCoinigy.Models
 {
     public class BalanceHistory
     {
-        public double Balance { get; set; }
+        public decimal Balance { get; set; }
         public DateTime DateTime { get; set; }
-        public double DollarAmount { get; set; }
+        public decimal DollarAmount { get; set; }
         public string Name { get; set; }
     }
 }
